+++ src/control/lib/spdk/include/nvme_control_internal.h
@@ -1,5 +1,5 @@
 /**
-* (C) Copyright 2019-2021 Intel Corporation.
+* (C) Copyright 2019-2022 Intel Corporation.
 *
 * SPDX-License-Identifier: BSD-2-Clause-Patent
 */
@@ -9,80 +9,7 @@
 
 #include <stdbool.h>
 #include <spdk/nvme_intel.h>
-
-#define BUFLEN 1024
-
-/**
- * \brief NVMECONTROL return codes
- */
-enum nvme_control_status_code {
-	NVMEC_SUCCESS			= 0x0,
-	NVMEC_ERR_CHK_SIZE		= 0x1,
-	NVMEC_ERR_GET_PCI_DEV		= 0x2,
-	NVMEC_ERR_PCI_ADDR_FMT		= 0x3,
-	NVMEC_ERR_PCI_ADDR_PARSE	= 0x4,
-	NVMEC_ERR_CTRLR_NOT_FOUND	= 0x5,
-	NVMEC_ERR_NS_NOT_FOUND		= 0x6,
-	NVMEC_ERR_NOT_SUPPORTED		= 0x7,
-	NVMEC_ERR_BAD_LBA		= 0x8,
-	NVMEC_ERR_ALLOC_IO_QPAIR	= 0x9,
-	NVMEC_ERR_NS_ID_UNEXPECTED	= 0xA,
-	NVMEC_ERR_NS_WRITE_FAIL		= 0xB,
-	NVMEC_ERR_MULTIPLE_ACTIVE_NS	= 0xC,
-	NVMEC_ERR_NULL_NS		= 0xD,
-	NVMEC_ERR_ALLOC_SEQUENCE_BUF	= 0xE,
-	NVMEC_ERR_NO_VMD_CTRLRS		= 0xF,
-	NVMEC_ERR_WRITE_TRUNC		= 0x10,
-	NVMEC_LAST_STATUS_VALUE
-};
-
-/**
- * \brief NVMe controller details
- */
-struct ctrlr_t {
-	char				 model[BUFLEN];
-	char				 serial[BUFLEN];
-	char				 pci_addr[BUFLEN];
-	char				 fw_rev[BUFLEN];
-	char				 pci_type[BUFLEN];
-	int				 socket_id;
-	struct ns_t			*nss;
-	struct nvme_stats		*stats;
-	struct ctrlr_t			*next;
-};
-
-/**
- * \brief NVMe namespace details
- */
-struct ns_t {
-	uint32_t	id;
-	uint64_t	size;
-	struct ns_t    *next;
-};
-
-/**
- * \brief Result struct for namespace wipe operation containing return code,
- * namespace id, parent controller pci address, info message and link to next
- * list element.
- */
-struct wipe_res_t {
-	char			 ctrlr_pci_addr[BUFLEN];
-	uint32_t		 ns_id;
-	int			 rc;
-	char			 info[BUFLEN];
-	struct wipe_res_t	*next;
-};
-
-/**
- * \brief Return containing return code, controllers, namespaces, wwipe
- * results and info message
- */
-struct ret_t {
-	struct ctrlr_t		*ctrlrs;
-	struct wipe_res_t	*wipe_results;
-	int			 rc;
-	char			 info[BUFLEN];
-};
+#include "nvme_control.h"
 
 struct ctrlr_entry {
 	struct spdk_nvme_ctrlr	*ctrlr;
@@ -149,15 +76,6 @@ struct ret_t *
 init_ret(void);
 
 /**
- * Free memory allocated in linked lists attached to the ret_t struct.
- *
- * \param ret A pointer to a return struct (ret_t) which itself needs to be
- *            freed explicitly after calling clean_ret.
- **/
-void
-clean_ret(struct ret_t *ret);
-
-/**
  * Get the NVMe controller
  *
  * \param centry (out) pointer to assign to pointer of found entry.
@@ -180,7 +98,7 @@ typedef int
 (*health_getter)(struct spdk_nvme_ctrlr *, struct health_entry *);
 
 struct ret_t *
-_discover(prober, bool, health_getter);
+_discover(prober, health_getter);
 
 /**
  * Provide ability to pass function pointers to _collect for mocking
@@ -205,14 +123,4 @@ _collect(struct ret_t *, data_copier, pci_getter, socket_id_getter);
  */
 struct ret_t *
 collect(void);
-
-/**
- * Cleanup allocated memory for controller list generated by probe/attach.
- *
- * \param detach flag to signify whether nvme controllers should be detached
- *               from SPDK during cleanup.
- */
-void
-cleanup(bool detach);
-
 #endif
