+++ src/control/lib/spdk/nvme.go
@@ -22,7 +22,6 @@ package spdk
 #include "spdk/env.h"
 #include "spdk/nvme.h"
 #include "include/nvme_control.h"
-#include "include/nvme_control_common.h"
 */
 import "C"
 
@@ -124,7 +123,10 @@ func (n *NvmeImpl) Discover(log logging.Logger) (storage.NvmeControllers, error)
 		return nil, errors.New("nil NvmeImpl")
 	}
 
-	ctrlrs, err := collectCtrlrs(C.nvme_discover(), "NVMe Discover(): C.nvme_discover")
+	retPtr := C.nvme_discover()
+	ctrlrs, err := collectCtrlrs(retPtr, "NVMe Discover(): C.nvme_discover")
+	C.nvme_clean_ret(retPtr)
+	C.nvme_clean_globals()
 
 	pciAddrs := ctrlrPCIAddresses(ctrlrs)
 	log.Debugf("discovered nvme ssds: %v", pciAddrs)
@@ -150,8 +152,10 @@ func (n *NvmeImpl) Format(log logging.Logger) ([]*FormatResult, error) {
 		return nil, errors.New("nil NvmeImpl")
 	}
 
-	results, err := collectFormatResults(C.nvme_wipe_namespaces(),
-		"NVMe Format(): C.nvme_wipe_namespaces()")
+	retPtr := C.nvme_wipe_namespaces()
+	results, err := collectFormatResults(retPtr, "NVMe Format(): C.nvme_wipe_namespaces()")
+	C.nvme_clean_ret(retPtr)
+	C.nvme_clean_globals()
 
 	pciAddrs := resultPCIAddresses(results)
 	log.Debugf("formatted nvme ssds: %v", pciAddrs)
@@ -251,12 +255,6 @@ func c2GoFormatResult(fmtResult *C.struct_wipe_res_t) *FormatResult {
 	}
 }
 
-// clean deallocates memory in return structure and frees the pointer.
-func clean(retPtr *C.struct_ret_t) {
-	C.clean_ret(retPtr)
-	C.free(unsafe.Pointer(retPtr))
-}
-
 // checkRet returns fault if ret_t struct input is nil or rc is non-zero.
 func checkRet(retPtr *C.struct_ret_t, msgFail string) error {
 	if retPtr == nil {
@@ -288,8 +286,6 @@ func checkRet(retPtr *C.struct_ret_t, msgFail string) error {
 
 // collectCtrlrs parses return struct to collect slice of nvme.Controller.
 func collectCtrlrs(retPtr *C.struct_ret_t, msgFail string) (storage.NvmeControllers, error) {
-	defer clean(retPtr)
-
 	if err := checkRet(retPtr, msgFail); err != nil {
 		return nil, err
 	}
@@ -324,8 +320,6 @@ func collectCtrlrs(retPtr *C.struct_ret_t, msgFail string) (storage.NvmeControll
 // collectFormatResults parses return struct to collect slice of
 // nvme.FormatResult.
 func collectFormatResults(retPtr *C.struct_ret_t, msgFail string) ([]*FormatResult, error) {
-	defer clean(retPtr)
-
 	if err := checkRet(retPtr, msgFail); err != nil {
 		return nil, err
 	}
